webportal_msb/requirements.txt
cryptography>=41.0.4
Flask>=3.0.0
flask-wtf>=1.2.1
flask_talisman
itsdangerous>=2.1.2
passlib>=1.7.4
argon2_cffi>=23.1.0
PyJWT>=2.8.0
pymongo>=4.6.0
requests>=2.31.0
Password_Validation==0.1.1
python-dotenv>=1.0.0
pyOpenSSL>=23.3.0
uwsgi
